refactor: move remark logic to separate files
