+++ components/article-list.tsx
@@ -10,7 +10,7 @@ export const ArticleList: React.FC<Props> = ({ articles }) => {
   return (
     <div className="articles">
       {articles.map((article) => (
-        <div key={article.id} className="article-name">
+        <div key={article.slug} className="article-name">
           <span className="date">{article.date}</span>
           <Link href={`/wiki/${article.slug}`}>
             <a style={{ marginLeft: '0.3rem' }}>{article.title}</a>
