+++ types/article.ts
@@ -1,6 +1,5 @@
 export interface Article {
   date: string
-  id: string
   slug: string
   title: string
 }
